Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
numpy>=1.26.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
email-validator>=2.0.0
reportlab>=4.0.0
